Fix: C4267

+++ code/glTF2Importer.cpp
@@ -454,7 +454,7 @@ void glTF2Importer::ImportMeshes(glTF2::Asset& r)
 
             std::vector<Mesh::Primitive::Target>& targets = prim.targets;
             if (targets.size() > 0) {
-                aim->mNumAnimMeshes = targets.size();
+                aim->mNumAnimMeshes = (unsigned int)targets.size();
                 aim->mAnimMeshes = new aiAnimMesh*[aim->mNumAnimMeshes];
                 for (size_t i = 0; i < targets.size(); i++) {
                     aim->mAnimMeshes[i] = aiCreateAnimMesh(aim);
